requirements.txt
openai
yt-dlp
pysrt
python-decouple
sounddevice
numpy
keyboard
python-dotenv
openai-whisper
openai